add filterEvents.js and filterByLangTech to vocabRequests.js
