+++ components/filterButtons.js
@@ -2,10 +2,10 @@ import renderToDom from '../utils/renderToDom';
 
 const filterButtons = () => {
   const domString = `
-  <button type="button" class="btn btn-dark btn-sm">HTML</button>
-  <button type="button" class="btn btn-dark btn-sm">CSS</button>
-  <button type="button" class="btn btn-dark btn-sm">JavaScript</button>
-  <button type="button" class="btn btn-dark btn-sm">REACT</button>
+  <button type="button" id="htmlFilter" class="btn btn-dark btn-sm">HTML</button>
+  <button type="button" id="cssFilter" class="btn btn-dark btn-sm">CSS</button>
+  <button type="button" id="javascriptFilter" class="btn btn-dark btn-sm">JavaScript</button>
+  <button type="button" id="reactFilter" class="btn btn-dark btn-sm">REACT</button>
   `;
 
   renderToDom('#filterButtonContainer', domString);
