--- components/vocabRequests.js
@@ -1,17 +1,0 @@
-import client from '../utils/client';
-
-const endpoint = client.databaseURL;
-
-const getVocabulary = () => new Promise((resolve, reject) => {
-  fetch(`${endpoint}/vocabulary.json`, {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  })
-    .then((response) => response.json())
-    .then((data) => resolve(Object.values(data)))
-    .catch(reject);
-});
-
-export default getVocabulary;
